Add config file support
Support yaml-based config file for nfd. Currently, this can only be used
for configuring the 'kernel' feature source. More specifically, for
specifying the kernel config file to use, and, the kconfig options that
are detected.

By default, nfd tries to read
/etc/kubernetes/node-feature-discovery/node-feature-discovery.conf.
This can be changed by specifying the --config command line flag.