Implement cpu feature source
Currently, it only discovers one feature label, i.e. 'logical-cpus',
which basically tells if hyper-threading has been enabled. However, the
label is not named 'hyper-threading' because the setting is not detected
directly, and, the heuristics can be mislead. NFD determines the thread
siblings of each (logical, online) cpu in the system. If any cpu has
thread siblings the feature is set to true. Thus, hyper-threading could
be effectively disabled by putting all sibling cpus offline (even if HT
would be enabled in hardware).