Install local-hooks into the NFD docker image
Install local feature detector hooks into the Docker image so that they
are automatically enabled, when building a custom image.

Change the NFD containers to run in privileged mode – this is required
in order for the cpu hook to be able to read MSRs. Also, add volume
mount in the NFD deployment templates to mount /dev/cpu inside the
container.