local-hooks: add 'memory' hook
This new custom hook implements detection of "die-clustering".
That it, it tries to determine if Cluster-on-Die or sub-NUMA clustering has
been enabled. These are technologies where a physical NUMA node is
further split into separate NUMA domains,

NOTE: It uses indirect guessing, based on counting the number of NUMA
nodes and physical CPU ids the kernel reports. Thus, the method used
here is inherently fragile. E.g. it can be fooled e.g. by Linux NUMA
emulation, and, it cannot be guaranteed that it will work correctly on
future CPU generations.