source/custom: add special name '*' for matching against all keys
Using '*' as the field name in an expressoin makes is possible to run
the expression agains all feature names, instead of the value of one
feature.

This can be useful in template rules for creating per-feature labels,
based on feature names (instead of values) and in "normal"
(non-template) rules for checking if (at least) one of certain feature
names are present.

An example of creating an "avx512" label if any AVX512* CPUID feature is
present:

  - name: "avx512-any-feature"
    labels: {avx512}
    matchFeatures:
      - feature: cpu.cpuid:
        matchExpressions:
          "*": {op: InRegexp, value: ["^AVX512"]}

An example of a template rule creating a dynamic set of labels  based on
the existence of kconfig options.

  - name: "kconfig-template-features"
    labelsTemplate: |
      {{range .kernel.config}}kconfig-{{.Name}}={{.Value}}
      {{end}}
    matchFeatures:
      - feature: kernel.config
        matchExpressions:
          "*": {op: In, value: [SWAP, X86, ARM]}