nfd-master: process labeling rules from CRs
Enable Custom Resource based label creation in nfd-master. This extends
the previously implemented controller stub for watching CRs. NFD-master
processes the rules on every incoming labeling request from workers.
Instead of node labels the functionality relies on the "raw features"
(identical to how nfd-worker handles custom rules) submitted by
nfd-worker, making it independent of the label source configuration of
the worker. This means that the labeling functions as expected even if
all sources in the worker are disabled.

NOTE: nfd-master is stateless and re-labeling is only triggered on
SetLabelsRequest from the worker – i.e. happening on intervals specified
by the core.sleepInterval configuration option (or -sleep-interval
cmdline flag). This means that modification/creation of the labelrule
CRs does not automatically update the node labels, but, the changes only
come visible when workers send their labeling requests.