pkg/apis/nfd: specify CRD for custom labelling rules
Add a Custom Resource Definition for specifying labelling rules. This
patch virtually replicates the source.custom.FeatureSpec in the CRD api
(located in the pkg/apis/nfd/v1alpha1 package) with the notable
exception that "MatchOn" legacy rules are not supported. Legacy rules
are left out in order to keep the CRD simple and clean.

This patch utilizes controller-gen (from sigs.k8s.io/controller-tools)
for generating the CRD and deepcopy methods. Code can be (re-)generated
with "make apigen". Install controller-gen with:

  go install sigs.k8s.io/controller-tools/cmd/controller-gen@v0.6.0

Update kustomize and helm deployments to deploy the CRD.